// traffic-one/functions/types/api.ts
export interface AuditLogAction {
  name: string
  metadata: Array<{ method?: string; route?: string; status?: number }>
}

export interface AuditLogActor {
  id: string
  type: string
  metadata: Array<{ email?: string; ip?: string; tokenType?: string }>
}

export interface AuditLogTarget {
  description: string
  metadata: Record<string, unknown>
}

export interface AuditLog {
  action: AuditLogAction
  actor: AuditLogActor
  target: AuditLogTarget
  occurred_at: string
}

export interface AuditLogsResponse {
  result: AuditLog[]
  retention_period: number
}
// ...
export interface MfaEnforcementResponse {
  enforced: boolean
}

export interface SSOProviderResponse {
  id: string
  organization_id: number
  enabled: boolean
  metadata_xml_file: string | null
  metadata_xml_url: string | null
  domains: string[]
  email_mapping: string[]
  first_name_mapping: string[]
  last_name_mapping: string[]
  user_name_mapping: string[]
  join_org_on_signup_enabled: boolean
  join_org_on_signup_role: string
  created_at: string
  updated_at: string
}

export interface CreateSSOProviderBody {
  enabled?: boolean
  metadata_xml_file?: string
  metadata_xml_url?: string
  domains?: string[]
  email_mapping?: string[]
  first_name_mapping?: string[]
  last_name_mapping?: string[]
  user_name_mapping?: string[]
  join_org_on_signup_enabled?: boolean
  join_org_on_signup_role?: string
}

export type UpdateSSOProviderBody = CreateSSOProviderBody